+++ crates/router/src/connector/utils.rs
@@ -10,6 +10,7 @@ use base64::Engine;
 use common_utils::{
     date_time,
     errors::ReportSwitchExt,
+    ext_traits::StringExt,
     pii::{self, Email, IpAddress},
 };
 use diesel_models::enums;
@@ -27,11 +28,13 @@ use crate::{
     consts,
     core::{
         errors::{self, ApiErrorResponse, CustomResult},
-        payments::{types::AuthenticationData, PaymentData, RecurringMandatePaymentData},
+        payments::{types::AuthenticationData, PaymentData},
     },
     pii::PeekInterface,
     types::{
-        self, api, domain, storage::enums as storage_enums, transformers::ForeignTryFrom,
+        self, api, domain,
+        storage::enums as storage_enums,
+        transformers::{ForeignFrom, ForeignTryFrom},
         ApplePayPredecryptData, BrowserInformation, PaymentsCancelData, ResponseId,
     },
     utils::{OptionExt, ValueExt},
@@ -86,7 +89,9 @@ pub trait RouterData {
     fn get_customer_id(&self) -> Result<String, Error>;
     fn get_connector_customer_id(&self) -> Result<String, Error>;
     fn get_preprocessing_id(&self) -> Result<String, Error>;
-    fn get_recurring_mandate_payment_data(&self) -> Result<RecurringMandatePaymentData, Error>;
+    fn get_recurring_mandate_payment_data(
+        &self,
+    ) -> Result<types::RecurringMandatePaymentData, Error>;
     #[cfg(feature = "payouts")]
     fn get_payout_method_data(&self) -> Result<api::PayoutMethodData, Error>;
     #[cfg(feature = "payouts")]
@@ -411,7 +416,9 @@ impl<Flow, Request, Response> RouterData for types::RouterData<Flow, Request, Re
             .to_owned()
             .ok_or_else(missing_field_err("preprocessing_id"))
     }
-    fn get_recurring_mandate_payment_data(&self) -> Result<RecurringMandatePaymentData, Error> {
+    fn get_recurring_mandate_payment_data(
+        &self,
+    ) -> Result<types::RecurringMandatePaymentData, Error> {
         self.recurring_mandate_payment_data
             .to_owned()
             .ok_or_else(missing_field_err("recurring_mandate_payment_data"))
@@ -1508,18 +1515,6 @@ impl AddressDetailsData for api::AddressDetails {
     }
 }
 
-pub trait BankRedirectBillingData {
-    fn get_billing_name(&self) -> Result<Secret<String>, Error>;
-}
-
-impl BankRedirectBillingData for domain::BankRedirectBilling {
-    fn get_billing_name(&self) -> Result<Secret<String>, Error> {
-        self.billing_name
-            .clone()
-            .ok_or_else(missing_field_err("billing_details.billing_name"))
-    }
-}
-
 pub trait MandateData {
     fn get_end_date(&self, format: date_time::DateFormat) -> Result<String, Error>;
     fn get_metadata(&self) -> Result<pii::SecretSerdeValue, Error>;
@@ -1545,7 +1540,7 @@ pub trait RecurringMandateData {
     fn get_original_payment_currency(&self) -> Result<enums::Currency, Error>;
 }
 
-impl RecurringMandateData for RecurringMandatePaymentData {
+impl RecurringMandateData for types::RecurringMandatePaymentData {
     fn get_original_payment_amount(&self) -> Result<i64, Error> {
         self.original_payment_authorized_amount
             .ok_or_else(missing_field_err("original_payment_authorized_amount"))
@@ -1795,72 +1790,80 @@ pub fn get_webhook_merchant_secret_key(connector_label: &str, merchant_id: &str)
 impl ForeignTryFrom<String> for UsStatesAbbreviation {
     type Error = error_stack::Report<errors::ConnectorError>;
     fn foreign_try_from(value: String) -> Result<Self, Self::Error> {
-        let binding = value.as_str().to_lowercase();
-        let state = binding.as_str();
-        match state {
-            "alabama" => Ok(Self::AL),
-            "alaska" => Ok(Self::AK),
-            "american samoa" => Ok(Self::AS),
-            "arizona" => Ok(Self::AZ),
-            "arkansas" => Ok(Self::AR),
-            "california" => Ok(Self::CA),
-            "colorado" => Ok(Self::CO),
-            "connecticut" => Ok(Self::CT),
-            "delaware" => Ok(Self::DE),
-            "district of columbia" | "columbia" => Ok(Self::DC),
-            "federated states of micronesia" | "micronesia" => Ok(Self::FM),
-            "florida" => Ok(Self::FL),
-            "georgia" => Ok(Self::GA),
-            "guam" => Ok(Self::GU),
-            "hawaii" => Ok(Self::HI),
-            "idaho" => Ok(Self::ID),
-            "illinois" => Ok(Self::IL),
-            "indiana" => Ok(Self::IN),
-            "iowa" => Ok(Self::IA),
-            "kansas" => Ok(Self::KS),
-            "kentucky" => Ok(Self::KY),
-            "louisiana" => Ok(Self::LA),
-            "maine" => Ok(Self::ME),
-            "marshall islands" => Ok(Self::MH),
-            "maryland" => Ok(Self::MD),
-            "massachusetts" => Ok(Self::MA),
-            "michigan" => Ok(Self::MI),
-            "minnesota" => Ok(Self::MN),
-            "mississippi" => Ok(Self::MS),
-            "missouri" => Ok(Self::MO),
-            "montana" => Ok(Self::MT),
-            "nebraska" => Ok(Self::NE),
-            "nevada" => Ok(Self::NV),
-            "new hampshire" => Ok(Self::NH),
-            "new jersey" => Ok(Self::NJ),
-            "new mexico" => Ok(Self::NM),
-            "new york" => Ok(Self::NY),
-            "north carolina" => Ok(Self::NC),
-            "north dakota" => Ok(Self::ND),
-            "northern mariana islands" => Ok(Self::MP),
-            "ohio" => Ok(Self::OH),
-            "oklahoma" => Ok(Self::OK),
-            "oregon" => Ok(Self::OR),
-            "palau" => Ok(Self::PW),
-            "pennsylvania" => Ok(Self::PA),
-            "puerto rico" => Ok(Self::PR),
-            "rhode island" => Ok(Self::RI),
-            "south carolina" => Ok(Self::SC),
-            "south dakota" => Ok(Self::SD),
-            "tennessee" => Ok(Self::TN),
-            "texas" => Ok(Self::TX),
-            "utah" => Ok(Self::UT),
-            "vermont" => Ok(Self::VT),
-            "virgin islands" => Ok(Self::VI),
-            "virginia" => Ok(Self::VA),
-            "washington" => Ok(Self::WA),
-            "west virginia" => Ok(Self::WV),
-            "wisconsin" => Ok(Self::WI),
-            "wyoming" => Ok(Self::WY),
-            _ => Err(errors::ConnectorError::InvalidDataFormat {
-                field_name: "address.state",
+        let state_abbreviation_check =
+            StringExt::<Self>::parse_enum(value.to_uppercase().clone(), "UsStatesAbbreviation");
+
+        match state_abbreviation_check {
+            Ok(state_abbreviation) => Ok(state_abbreviation),
+            Err(_) => {
+                let binding = value.as_str().to_lowercase();
+                let state = binding.as_str();
+                match state {
+                    "alabama" => Ok(Self::AL),
+                    "alaska" => Ok(Self::AK),
+                    "american samoa" => Ok(Self::AS),
+                    "arizona" => Ok(Self::AZ),
+                    "arkansas" => Ok(Self::AR),
+                    "california" => Ok(Self::CA),
+                    "colorado" => Ok(Self::CO),
+                    "connecticut" => Ok(Self::CT),
+                    "delaware" => Ok(Self::DE),
+                    "district of columbia" | "columbia" => Ok(Self::DC),
+                    "federated states of micronesia" | "micronesia" => Ok(Self::FM),
+                    "florida" => Ok(Self::FL),
+                    "georgia" => Ok(Self::GA),
+                    "guam" => Ok(Self::GU),
+                    "hawaii" => Ok(Self::HI),
+                    "idaho" => Ok(Self::ID),
+                    "illinois" => Ok(Self::IL),
+                    "indiana" => Ok(Self::IN),
+                    "iowa" => Ok(Self::IA),
+                    "kansas" => Ok(Self::KS),
+                    "kentucky" => Ok(Self::KY),
+                    "louisiana" => Ok(Self::LA),
+                    "maine" => Ok(Self::ME),
+                    "marshall islands" => Ok(Self::MH),
+                    "maryland" => Ok(Self::MD),
+                    "massachusetts" => Ok(Self::MA),
+                    "michigan" => Ok(Self::MI),
+                    "minnesota" => Ok(Self::MN),
+                    "mississippi" => Ok(Self::MS),
+                    "missouri" => Ok(Self::MO),
+                    "montana" => Ok(Self::MT),
+                    "nebraska" => Ok(Self::NE),
+                    "nevada" => Ok(Self::NV),
+                    "new hampshire" => Ok(Self::NH),
+                    "new jersey" => Ok(Self::NJ),
+                    "new mexico" => Ok(Self::NM),
+                    "new york" => Ok(Self::NY),
+                    "north carolina" => Ok(Self::NC),
+                    "north dakota" => Ok(Self::ND),
+                    "northern mariana islands" => Ok(Self::MP),
+                    "ohio" => Ok(Self::OH),
+                    "oklahoma" => Ok(Self::OK),
+                    "oregon" => Ok(Self::OR),
+                    "palau" => Ok(Self::PW),
+                    "pennsylvania" => Ok(Self::PA),
+                    "puerto rico" => Ok(Self::PR),
+                    "rhode island" => Ok(Self::RI),
+                    "south carolina" => Ok(Self::SC),
+                    "south dakota" => Ok(Self::SD),
+                    "tennessee" => Ok(Self::TN),
+                    "texas" => Ok(Self::TX),
+                    "utah" => Ok(Self::UT),
+                    "vermont" => Ok(Self::VT),
+                    "virgin islands" => Ok(Self::VI),
+                    "virginia" => Ok(Self::VA),
+                    "washington" => Ok(Self::WA),
+                    "west virginia" => Ok(Self::WV),
+                    "wisconsin" => Ok(Self::WI),
+                    "wyoming" => Ok(Self::WY),
+                    _ => Err(errors::ConnectorError::InvalidDataFormat {
+                        field_name: "address.state",
+                    }
+                    .into()),
+                }
             }
-            .into()),
         }
     }
 }
@@ -1868,26 +1871,33 @@ impl ForeignTryFrom<String> for UsStatesAbbreviation {
 impl ForeignTryFrom<String> for CanadaStatesAbbreviation {
     type Error = error_stack::Report<errors::ConnectorError>;
     fn foreign_try_from(value: String) -> Result<Self, Self::Error> {
-        let binding = value.as_str().to_lowercase();
-        let state = binding.as_str();
-        match state {
-            "alberta" => Ok(Self::AB),
-            "british columbia" => Ok(Self::BC),
-            "manitoba" => Ok(Self::MB),
-            "new brunswick" => Ok(Self::NB),
-            "newfoundland and labrador" | "newfoundland & labrador" => Ok(Self::NL),
-            "northwest territories" => Ok(Self::NT),
-            "nova scotia" => Ok(Self::NS),
-            "nunavut" => Ok(Self::NU),
-            "ontario" => Ok(Self::ON),
-            "prince edward island" => Ok(Self::PE),
-            "quebec" => Ok(Self::QC),
-            "saskatchewan" => Ok(Self::SK),
-            "yukon" => Ok(Self::YT),
-            _ => Err(errors::ConnectorError::InvalidDataFormat {
-                field_name: "address.state",
+        let state_abbreviation_check =
+            StringExt::<Self>::parse_enum(value.to_uppercase().clone(), "CanadaStatesAbbreviation");
+        match state_abbreviation_check {
+            Ok(state_abbreviation) => Ok(state_abbreviation),
+            Err(_) => {
+                let binding = value.as_str().to_lowercase();
+                let state = binding.as_str();
+                match state {
+                    "alberta" => Ok(Self::AB),
+                    "british columbia" => Ok(Self::BC),
+                    "manitoba" => Ok(Self::MB),
+                    "new brunswick" => Ok(Self::NB),
+                    "newfoundland and labrador" | "newfoundland & labrador" => Ok(Self::NL),
+                    "northwest territories" => Ok(Self::NT),
+                    "nova scotia" => Ok(Self::NS),
+                    "nunavut" => Ok(Self::NU),
+                    "ontario" => Ok(Self::ON),
+                    "prince edward island" => Ok(Self::PE),
+                    "quebec" => Ok(Self::QC),
+                    "saskatchewan" => Ok(Self::SK),
+                    "yukon" => Ok(Self::YT),
+                    _ => Err(errors::ConnectorError::InvalidDataFormat {
+                        field_name: "address.state",
+                    }
+                    .into()),
+                }
             }
-            .into()),
         }
     }
 }
@@ -2181,7 +2191,7 @@ pub fn is_refund_failure(status: enums::RefundStatus) -> bool {
 }
 
 impl
-    From<(
+    ForeignFrom<(
         Option<String>,
         Option<String>,
         Option<String>,
@@ -2190,7 +2200,7 @@ impl
         Option<String>,
     )> for types::ErrorResponse
 {
-    fn from(
+    fn foreign_from(
         (code, message, reason, http_code, attempt_status, connector_transaction_id): (
             Option<String>,
             Option<String>,